eric-oss-core-reporting-dimension-query-security/src/test/java/com/ericsson/oss/air/security/RestTemplateReloaderTest.java
/*******************************************************************************
 * COPYRIGHT Ericsson 2023
 *
 *
 *
 * The copyright to the computer program(s) herein is the property of
 *
 * Ericsson Inc. The programs may be used and/or copied only with written
 *
 * permission from Ericsson Inc. or in accordance with the terms and
 *
 * conditions stipulated in the agreement/contract under which the
 *
 * program(s) have been supplied.
 ******************************************************************************/

package com.ericsson.oss.air.security;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.Spy;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.boot.ssl.SslBundle;
import org.springframework.boot.ssl.SslBundles;
import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
import org.springframework.http.client.SimpleClientHttpRequestFactory;
import org.springframework.web.client.RestTemplate;

import static org.assertj.core.api.Assertions.assertThat;
import static org.mockito.ArgumentMatchers.anyString;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
public class RestTemplateReloaderTest {

    @InjectMocks
    private RestTemplateReloader restTemplateReloader;

    @Spy
    private RestTemplate restTemplate;

    @Mock
    private SslBundles sslBundles;

    @Test
    void testRestTemplateReloader() {
        when(sslBundles.getBundle(anyString())).thenReturn(SslBundle.of(new CustomSslStoreBundle()));
        assertThat(restTemplate.getRequestFactory()).isInstanceOf(SimpleClientHttpRequestFactory.class);
        restTemplateReloader.reload();
        assertThat(restTemplate.getRequestFactory()).isNotNull();
        assertThat(restTemplate.getRequestFactory()).isInstanceOf(HttpComponentsClientHttpRequestFactory.class);
    }

    @Test
    void testRestTemplateReloaderWithCtsTlsDisabled() {
        restTemplateReloader.reload();
        assertThat(restTemplate.getRequestFactory()).isNotNull();
        assertThat(restTemplate.getRequestFactory()).isInstanceOf(SimpleClientHttpRequestFactory.class);
    }
}
